backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
PyGithub
google-generativeai
flake8
openai
anthropic
docker
requests
python-jose[cryptography]